Move roslyn rule and options descriptions from tests
